requirements.txt
fastapi==0.115.12
nltk==3.9.1
numpy==2.3.0
openai==1.86.0
psycopg2_binary==2.9.10
pydantic==2.11.5
selenium==4.33.0
sentence_transformers==3.4.1
undetected_chromedriver==3.5.5